fix: do not output logs to stderr
Standard error is seen as an application error by Log Analytics and
counted as a failure.

According to https://github.com/PowerShell/PowerShell/issues/10476
this should be the right workaround to redirect stderr to stdout
when using Invoke-Expression.

Also log every second to have a bit more information even when there
is nothing to do.